// source/mission-management-frontend/src/lib/domains/mission/models/RelatedPlaceRefOrValue.ts
export interface RelatedPlaceRefOrValue {
    /**
     * Unique identifier of the place
     * @type {string}
     * @memberof RelatedPlaceRefOrValue
     */
    id?: string;
    /**
     * Unique reference of the place
     * @type {string}
     * @memberof RelatedPlaceRefOrValue
     */
    href?: string;
    /**
     * A user-friendly name for the place, such as [Paris Store], [London Store], [Main Home]
     * @type {string}
     * @memberof RelatedPlaceRefOrValue
     */
    name?: string;
    /**
     *
     * @type {string}
     * @memberof RelatedPlaceRefOrValue
     */
    role: string;
    /**
     * When sub-classing, this defines the super-class
     * @type {string}
     * @memberof RelatedPlaceRefOrValue
     */
    baseType?: string;
    /**
     * A URI to a JSON-Schema file that defines additional attributes and relationships
     * @type {string}
     * @memberof RelatedPlaceRefOrValue
     */
    schemaLocation?: string;
    /**
     * When sub-classing, this defines the sub-class Extensible name
     * @type {string}
     * @memberof RelatedPlaceRefOrValue
     */
    type?: string;
    /**
     * The actual type of the target instance when needed for disambiguation.
     * @type {string}
     * @memberof RelatedPlaceRefOrValue
     */
    referredType?: string;
}